fix: shape not 'activated' if pen is not in use.
Previously, when selecting a shape (say an arrow) while using a tool other
than the pen (say an eraser), xournalpp still remains with the eraser.
Also, this is not mitigated by selecting the pen after selecting a shape,
as selecting the will clearShape().

The remedy is to first clickPen(), followed by click<shape>().